PlayVisionV1/backend/requirements.txt
Django==5.2.8
djangorestframework==3.16.1
Pillow==12.0.0
psycopg2-binary==2.9.6
django-cors-headers==4.9.0
django-seed== 0.3.1
faker==40.1.0
gunicorn==23.0.0
pytest == 9.0.2
pytest-django == 4.11.1
pytest-cov == 7.0.0
factory_boy == 3.3.3